Implement coding standards documentation, add VSCode settings, create mappers for AudioFile and AudioPreprocessing, update EF Core migration files for audio file models, and adjust database schema for audio duration and file size.

+++ VoiceByAuribus.API/Features/AudioFiles/Application/Dtos/AudioPreprocessingResponseDto.cs
@@ -8,7 +8,7 @@ namespace VoiceByAuribus_API.Features.AudioFiles.Application.Dtos;
 public class AudioPreprocessingResponseDto
 {
     public required string Status { get; set; }
-    public decimal? AudioDurationSeconds { get; set; }
+    public int? AudioDurationSeconds { get; set; }
     public string? S3UriShort { get; set; }
     public string? S3UriInference { get; set; }
     public DateTime? ProcessingStartedAt { get; set; }
